+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 
+## 0.6.0
+  - **[NEW]** Implementation of create Record
+  - **[Example]** of create Tickets in Customer Mode
+  
 ## 0.5.0
-  - Implementation of RedooWeekplanner Module
+  - **[NEW]** Implementation of RedooWeekplanner Module
   
 ## 0.4.0
   - First implementation